Add files via upload
Correlation is a technique that is often employed both in the field of Statistics and machine learning to evaluate the strength and direction of association between two variables and more importantly to test for multicollinearity before performing regression analysis. The following github link provides six simple steps to computing correlation coefficient and plotting the correlation matrix.